Move MiqFaultTolerantVim to VMwareWebService
