+++ lib/gems/pending/VMwareWebService/miq_fault_tolerant_vim.rb
